+++ gulpfile.js
@@ -3,6 +3,8 @@ const mergeStream = require('merge-stream')
 const sass = require('sass')
 const yupSass = require('gulp-sass')(sass)
 const gulpZip = require('gulp-zip')
+const tar = require('gulp-tar')
+const gzip = require('gulp-gzip')
 const pkg = require('pkg')
 const path = require('path')
 const del = require('del')
@@ -10,6 +12,7 @@ const minify = require('gulp-minify')
 const concat = require('gulp-concat')
 const sort = require('gulp-sort')
 const rename = require('gulp-rename')
+const hashsum = require('gulp-hashsum')
 const resedit = require('resedit/cjs')
 const fs = require('fs')
 
@@ -170,6 +173,7 @@ const buildSafari = () => {
     src(getWebPath('./_locales/**/*')).pipe(dest(localesDir)),
     // src(getWebPath('./libs/**/*')).pipe(dest(libsDir)),
     src(getWebPath('./libs/**/*'))
+      .pipe(sort())
       .pipe(concat('libs.min.js'))
       .pipe(dest(libsDir)),
     src(getCorePath('./fonts/**/*')).pipe(dest(fontsDir)),
@@ -282,14 +286,30 @@ const setWindowsExeInfo = async () => {
 
 const zipMacOSX64 = () => {
   return src(distUtils.getFilePath('MACOS_X64', true))
+    .pipe(rename(distUtils.getFileNameOriginal('MACOS_X64_FRIENDLY')))
     .pipe(gulpZip(distUtils.getFileNameZip('MACOS_X64')))
     .pipe(dest(distUtils.getFileDir('MACOS_X64')))
 }
 
 const zipMacOSARM64 = () => {
   return src(distUtils.getFilePath('MACOS_ARM64', true))
+    .pipe(rename(distUtils.getFileNameOriginal('MACOS_ARM64_FRIENDLY')))
     .pipe(gulpZip(distUtils.getFileNameZip('MACOS_ARM64')))
     .pipe(dest(distUtils.getFileDir('MACOS_ARM64')))
+}
+
+const tarGzipMacOSX64 = () => {
+  return src(distUtils.getFilePath('MACOS_X64', true))
+    .pipe(rename('zadark'))
+    .pipe(tar(distUtils.getFileNameMacOSTar()))
+    .pipe(gzip())
+    .pipe(dest(distUtils.getFileDir('MACOS_X64')))
+}
+
+const hashsumMacOS = () => {
+  const inp = path.join(distUtils.getFileDir('MACOS_X64'), distUtils.getFileNameMacOSTar() + '.gz')
+  const out = distUtils.getFileDir('MACOS_X64')
+  return src(inp).pipe(hashsum({ hash: 'sha256', dest: out }))
 }
 
 const zipWindows = () => {
@@ -324,13 +344,26 @@ const edgeDist = () => {
     .pipe(dest(distUtils.getFileDir('EDGE')))
 }
 
+const delTmpFile = (patterns) => {
+  return function delTmpFile () {
+    return del(patterns)
+  }
+}
+
 const pcDist = series(
   pkgMacOS,
   pkgWindows,
   setWindowsExeInfo,
   zipMacOSX64,
   zipMacOSARM64,
-  zipWindows
+  tarGzipMacOSX64,
+  hashsumMacOS,
+  zipWindows,
+  parallel(
+    delTmpFile(distUtils.getFilePath('MACOS_X64', true)),
+    delTmpFile(distUtils.getFilePath('MACOS_ARM64', true)),
+    delTmpFile(distUtils.getFilePath('WINDOWS', true))
+  )
 )
 
 // Exports
@@ -361,6 +394,23 @@ const distAll = series(
   )
 )
 
+const distWeb = series(
+  buildAll,
+  parallel(
+    chromeDist,
+    firefoxDist,
+    operaDist,
+    edgeDist
+  )
+)
+
+const distPC = series(
+  buildAll,
+  parallel(
+    pcDist
+  )
+)
+
 const watchAll = () => {
   watch([
     getCorePath('./**/*'),
@@ -376,4 +426,6 @@ const dev = series(buildAll, watchAll)
 
 exports.build = buildAll
 exports.dist = distAll
+exports.distWeb = distWeb
+exports.distPC = distPC
 exports.dev = dev